Allow for more than 256 occurrences of an argument.

Hi,

I often use tools with `xargs` and the like, and it's entirely possible to end up with inhumanly long sets of parameters being passed to a command.

Without this fix, the library will panic with an arithmetic overflow in `ArgMatcher#inc_occurrence_of` when the 256th instance of the parameter is accounted for.

Thanks,

+++ src/args/matched_arg.rs
@@ -6,7 +6,7 @@ use vec_map::VecMap;
 #[derive(Debug, Clone)]
 pub struct MatchedArg {
     #[doc(hidden)]
-    pub occurs: u8,
+    pub occurs: u64,
     #[doc(hidden)]
     pub vals: VecMap<OsString>,
 }
